trouble with database
I cant seem to commit my POST commands to m db.json file. I've managed to link the db.json to the html so that it is referenced in the browser as being a part of the html, but it's still not working. I need to figure out why it's not pushing changes to the db.json. Then I need to figure out the delete function and I should be complete