things are getting weird
but im almost done. I made a formula to turn the id to and from integers so they can be processed outside mysql using my method. I almost got everything working. The only think I cant figure out, when I try to delete #3 out of 3, it deletes 1 and 3, leaving only 2 I have a theory on that