im stuck
Im getting the data to render the way I want in the console log, now why wont it write that way? I've tried all express write functions, which should work to overwrite the file with my data, but instead it returns strange results. It seems to be deleting all notes except for the note directly preceding it, or other functions leave no data altogether. Maybe a function is being run twice somehow? If I delete note #3 it deletes note #3 and #1. If I delete note #2 it deletes all notes but #1. If I delete note #1 it deletes all notes.